delete cast

+++ srcs/server/HTTPResponse.cpp
@@ -114,9 +114,9 @@ std::string HTTPResponse::GenerateDefaultHTML() const
 	std::stringstream ss;
 
 	ss << "<html>\r\n";
-	ss << "<head><title>" << static_cast<int>(status_code_) << " " << kStatusMsg_[status_code_] <<"</title></head>\r\n";
+	ss << "<head><title>" << status_code_ << " " << kStatusMsg_[status_code_] <<"</title></head>\r\n";
 	ss << "<body>\r\n";
-	ss << "<center><h1>" << static_cast<int>(status_code_) << " " << kStatusMsg_[status_code_] << "</h1></center>\r\n";
+	ss << "<center><h1>" << status_code_ << " " << kStatusMsg_[status_code_] << "</h1></center>\r\n";
 	ss << "<hr><center>" << "Webserv" << "</center>\r\n";
 	ss << "</body>\r\n";
 	ss << "</html>\r\n";
@@ -127,7 +127,7 @@ std::string HTTPResponse::CreateResponse()
 {
 	std::stringstream ss;
 
-	ss << "HTTP/1.1 " << static_cast<int>(status_code_) << " " << kStatusMsg_[status_code_] << "\r\n";
+	ss << "HTTP/1.1 " << status_code_ << " " << kStatusMsg_[status_code_] << "\r\n";
 	ss << HeaderFeild();
 	ss << body_;
 	return (ss.str());
